refactor: extract LoggingHelpers and DockerRunner from Program.cs
